# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/finance/defaults": {
      "get": {
        "operationId": "FinanceController_getDefaultsNew",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/v1/finance/defaults/{companyId}": {
      "get": {
        "operationId": "FinanceController_getDefaults",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
...
    "/v1/finance/sale-invoice": {
      "post": {
        "operationId": "FinanceController_createSaleInvoice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/v1/finance/sale-invoices": {
      "get": {
        "operationId": "FinanceController_listSaleInvoices",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/v1/finance/sale-invoices/{id}": {
      "get": {
        "operationId": "FinanceController_getSaleInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/v1/finance/payment-receipt": {
      "post": {
        "operationId": "FinanceController_addPaymentReceipt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentReceiptDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/v1/finance/pos-sale": {
      "post": {
        "operationId": "FinanceController_posSale",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePosSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
...
    "/v1/finance/purchase-invoices/{id}": {
      "get": {
        "operationId": "FinanceController_getPurchaseInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
...
    "/v1/finance/treasury/transactions": {
      "get": {
        "operationId": "FinanceController_treasuryTransactions",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Finance"
        ]
      }
    },
    "/v1/inventory/products": {
      "get": {
        "operationId": "InventoryController_listProducts",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      },
      "post": {
        "operationId": "InventoryController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/v1/inventory/products/{id}": {
      "get": {
        "operationId": "InventoryController_getProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      },
      "patch": {
        "operationId": "InventoryController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      },
      "delete": {
        "operationId": "InventoryController_deleteProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/v1/inventory/categories": {
      "get": {
        "operationId": "InventoryController_listCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      },
      "post": {
        "operationId": "InventoryController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/v1/inventory/categories/{id}": {
      "patch": {
        "operationId": "InventoryController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/v1/inventory/search": {
      "get": {
        "operationId": "InventoryController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "CorePOS Backend API",
    "description": "Backend migration service from Supabase to NestJS",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSaleInvoiceItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "uuid-product"
          },
          "qty": {
            "type": "number",
            "example": 1
          },
          "unitPrice": {
            "type": "number",
            "example": 10
          },
          "totalLine": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "productId",
          "qty",
          "unitPrice",
          "totalLine"
        ]
      },
      "CreateSaleInvoiceDto": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "string",
            "example": "uuid-branch"
          },
          "warehouseId": {
            "type": "string",
            "example": "uuid-warehouse"
          },
          "customerId": {
            "type": "object",
            "example": "uuid-customer",
            "nullable": true
          },
          "cashierId": {
            "type": "string",
            "example": "uuid-cashier"
          },
          "subtotal": {
            "type": "number",
            "example": 10
          },
          "discountAmount": {
            "type": "number",
            "example": 0
          },
          "taxAmount": {
            "type": "number",
            "example": 0
          },
          "total": {
            "type": "number",
            "example": 10
          },
          "paid": {
            "type": "number",
            "example": 10
          },
          "remaining": {
            "type": "number",
            "example": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSaleInvoiceItemDto"
            }
          }
        },
        "required": [
          "branchId",
          "warehouseId",
          "cashierId",
          "subtotal",
          "discountAmount",
          "taxAmount",
          "total",
          "paid",
          "remaining",
          "items"
        ]
      },
      "PaymentReceiptDto": {
        "type": "object",
        "properties": {
          "treasuryId": {
            "type": "string",
            "example": "uuid-treasury"
          },
          "amount": {
            "type": "number",
            "example": 5
          },
          "method": {
            "type": "string",
            "enum": [
              "cash",
              "card",
              "bank"
            ]
          },
          "notes": {
            "type": "string"
          },
          "invoiceId": {
            "type": "string",
            "description": "Optional link to a sale invoice to update paid/remaining and customer balance."
          },
          "customerId": {
            "type": "string",
            "description": "Optional customer id (required if invoiceId is not provided)."
          },
          "createdBy": {
            "type": "string",
            "example": "uuid-user"
          }
        },
        "required": [
          "treasuryId",
          "amount",
          "method",
          "createdBy"
        ]
      },
      "InvoiceLineDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "uuid-product"
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "unitPrice": {
            "type": "number",
            "example": 150
          }
        },
        "required": [
          "productId",
          "quantity",
          "unitPrice"
        ]
      },
      "CreatePosSaleDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "example": "uuid-company"
          },
          "branchId": {
            "type": "string",
            "example": "uuid-branch"
          },
          "warehouseId": {
            "type": "string",
            "example": "uuid-warehouse"
          },
          "treasuryId": {
            "type": "object",
            "example": "uuid-treasury",
            "nullable": true
          },
          "customerId": {
            "type": "object",
            "example": "uuid-customer",
            "nullable": true
          },
          "discountAmount": {
            "type": "number",
            "example": 0
          },
          "taxAmount": {
            "type": "number",
            "example": 0
          },
          "totalAmount": {
            "type": "number",
            "example": 300
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "cash",
              "card",
              "deferred"
            ]
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineDto"
            }
          }
        },
        "required": [
          "companyId",
          "branchId",
          "warehouseId",
          "discountAmount",
          "taxAmount",
          "totalAmount",
          "paymentMethod",
          "lines"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ماوس لاسلكي"
          },
          "nameEn": {
            "type": "string",
            "example": "Wireless Mouse"
          },
          "categoryId": {
            "type": "string",
            "example": "uuid-category"
          },
          "unitId": {
            "type": "string",
            "example": "uuid-unit"
          },
          "barcode": {
            "type": "string",
            "example": "1234567890"
          },
          "sku": {
            "type": "string",
            "example": "SKU-001"
          },
          "price1": {
            "type": "string",
            "example": "150",
            "description": "Numeric string for DB compatibility"
          },
          "costPrice": {
            "type": "string",
            "example": "100",
            "description": "Numeric string for DB compatibility"
          },
          "warehouseId": {
            "type": "string",
            "example": "uuid-warehouse"
          },
          "initialQty": {
            "type": "string",
            "example": "10",
            "description": "Numeric string for DB compatibility"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "إلكترونيات"
          },
          "parentId": {
            "type": "string",
            "example": "uuid-parent"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}